Move DataGraph into src/graphs
