+++ src/components/graphs/README.md
@@ -1,0 +1,4 @@
+# Testing issues
+
+`DataGraph` appears to pose some difficulties in testing.
+As it is a legacy component, it's hard to know
