+++ src/components/graphs/TimeSeriesGraph/TimeSeriesGraph.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Row, Col, ControlLabel } from 'react-bootstrap';
 
-import DataGraph from '../../DataGraph/DataGraph';
+import DataGraph from '../DataGraph/DataGraph';
 import styles from './TimeSeriesGraph.css';
 import {
   validateAnnualCycleData,
